pkg/kepval: Move directory validation functionality to package
